backend/requirements.txt
asgiref==3.11.0
boto3==1.42.39
botocore==1.42.39
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
jmespath==1.1.0
pillow==12.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
